% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasis.R
\docType{package}
\name{nasis-package}
\alias{nasis}
\alias{nasis-package}
\title{nasis: National Soil Information System (NASIS) Data Model Interface}
\usage{
nasis(dsn = NULL, ...)
}
\arguments{
\item{dsn}{A \code{DBIConnection} or path to SQLite database. Default (\code{NULL}) is equivalent to \code{soilDB::NASIS()}}

\item{...}{Additional arguments to \code{\link[dm:dm_from_con]{dm::dm_from_con()}}}
}
\value{
A \code{DBIConnection}
}
\description{
A 'NASIS' Data Model abstraction layer provided by the 'dm' package.

The function \code{nasis()} provides a wrapper for passing \code{\link[soilDB:dbConnectNASIS]{soilDB::NASIS()}}-like \code{DBIConnection} to \code{\link[dm:dm_from_con]{dm::dm_from_con()}} to create a \code{dm} object.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
